reorg spdm_device_secret_lib_sample
The original code is hard to maintain because we put so many features in one lib.c.

split lib.c to multiple smaller files.
rename cert.c to read_pub_cert.c

Only code move, zero code change.
