+++ os_stub/spdm_device_secret_lib_sample/CMakeLists.txt
@@ -12,8 +12,19 @@ target_include_directories(spdm_device_secret_lib_sample
 
 target_sources(spdm_device_secret_lib_sample
     PRIVATE
-        lib.c
-        cert.c
+        chal.c
+        csr.c
+        event.c
+        key_pair.c
+        meas.c
+        psk.c
+        read_priv_key_pem.c
+        read_priv_key_raw_data.c
+        read_pub_cert.c
+        read_pub_key_der.c
+        read_special_cert.c
+        set_cert.c
+        sign.c
 )
 
 if ((ARCH STREQUAL "arm") OR (ARCH STREQUAL "aarch64"))
