+++ os_stub/spdm_device_secret_lib_sample/read_pub_cert.c
@@ -1,12 +1,7 @@
 /**
  *  Copyright Notice:
- *  Copyright 2021-2022 DMTF. All rights reserved.
+ *  Copyright 2021-2024 DMTF. All rights reserved.
  *  License: BSD 3-Clause License. For full text see link: https://github.com/DMTF/libspdm/blob/main/LICENSE.md
- **/
-
-/** @file
- * SPDM common library.
- * It follows the SPDM Specification.
  **/
 
 #include <stdarg.h>
@@ -21,6 +16,7 @@
 #include <base.h>
 #include "library/memlib.h"
 #include "spdm_device_secret_lib_internal.h"
+#include "internal/libspdm_common_lib.h"
 
 bool libspdm_read_responder_root_public_certificate(uint32_t base_hash_algo,
                                                     uint32_t base_asym_algo,
@@ -1087,187 +1083,47 @@ bool libspdm_read_requester_public_certificate_chain(
     return true;
 }
 
-bool libspdm_read_responder_root_public_certificate_by_size(
-    uint32_t base_hash_algo, uint32_t base_asym_algo, uint16_t chain_id,
-    void **data, size_t *size, void **hash, size_t *hash_size)
+bool libspdm_read_responder_certificate(uint32_t base_asym_algo,
+                                        void **data, size_t *size)
 {
     bool res;
-    void *file_data;
-    size_t file_size;
-    spdm_cert_chain_t *cert_chain;
-    size_t cert_chain_size;
     char *file;
-    size_t digest_size;
 
-    *data = NULL;
-    *size = 0;
-    if (hash != NULL) {
-        *hash = NULL;
-    }
-    if (hash_size != NULL) {
-        *hash_size = 0;
-    }
-
-    switch (chain_id) {
-    case LIBSPDM_TEST_CERT_SMALL:
-        file = "long_chains/Shorter1024B_ca.cert.der";
+    switch (base_asym_algo) {
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_RSASSA_2048:
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_RSAPSS_2048:
+        file = "rsa2048/end_responder.cert.der";
         break;
-    case LIBSPDM_TEST_CERT_MAXINT16: /* data_size slightly smaller than 0x7FFF*/
-        file = "long_chains/ShorterMAXINT16_ca.cert.der";
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_RSASSA_3072:
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_RSAPSS_3072:
+        file = "rsa3072/end_responder.cert.der";
         break;
-    case LIBSPDM_TEST_CERT_MAXUINT16: /* data_size slightly smaller than 0xFFFF*/
-        file = "long_chains/ShorterMAXUINT16_ca.cert.der";
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_RSASSA_4096:
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_RSAPSS_4096:
+        file = "rsa4096/end_responder.cert.der";
         break;
-    case LIBSPDM_LIBSPDM_TEST_CERT_MAXUINT16_LARGER: /* data_size larger than 0xFFFF*/
-        file = "long_chains/LongerMAXUINT16_ca.cert.der";
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_ECDSA_ECC_NIST_P256:
+        file = "ecp256/end_responder.cert.der";
+        break;
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_ECDSA_ECC_NIST_P384:
+        file = "ecp384/end_responder.cert.der";
+        break;
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_ECDSA_ECC_NIST_P521:
+        file = "ecp521/end_responder.cert.der";
+        break;
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_TPM_ALG_SM2_ECC_SM2_P256:
+        file = "sm2/end_responder.cert.der";
+        break;
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_EDDSA_ED25519:
+        file = "ed25519/end_responder.cert.der";
+        break;
+    case SPDM_ALGORITHMS_BASE_ASYM_ALGO_EDDSA_ED448:
+        file = "ed448/end_responder.cert.der";
         break;
     default:
         LIBSPDM_ASSERT(false);
         return false;
     }
-    res = libspdm_read_input_file(file, &file_data, &file_size);
-    if (!res) {
-        return res;
-    }
-
-    digest_size = libspdm_get_hash_size(base_hash_algo);
-
-    cert_chain_size = sizeof(spdm_cert_chain_t) + digest_size + file_size;
-    cert_chain = (void *)malloc(cert_chain_size);
-    if (cert_chain == NULL) {
-        free(file_data);
-        return false;
-    }
-    cert_chain->length = (uint16_t)cert_chain_size;
-    cert_chain->reserved = 0;
-
-    res = libspdm_hash_all(base_hash_algo, file_data, file_size,
-                           (uint8_t *)(cert_chain + 1));
-    if (!res) {
-        free(file_data);
-        free(cert_chain);
-        return res;
-    }
-    libspdm_copy_mem((uint8_t *)cert_chain + sizeof(spdm_cert_chain_t) + digest_size,
-                     cert_chain_size - (sizeof(spdm_cert_chain_t) + digest_size),
-                     file_data, file_size);
-
-    *data = cert_chain;
-    *size = cert_chain_size;
-    if (hash != NULL) {
-        *hash = (cert_chain + 1);
-    }
-    if (hash_size != NULL) {
-        *hash_size = digest_size;
-    }
-
-    free(file_data);
-    return true;
-}
-
-bool libspdm_read_responder_public_certificate_chain_by_size(
-    uint32_t base_hash_algo, uint32_t base_asym_algo, uint16_t chain_id,
-    void **data, size_t *size, void **hash, size_t *hash_size)
-{
-    bool res;
-    void *file_data;
-    size_t file_size;
-    spdm_cert_chain_t *cert_chain;
-    size_t cert_chain_size;
-    char *file;
-    const uint8_t *root_cert;
-    size_t root_cert_len;
-    size_t digest_size;
-    bool is_requester_cert;
-    bool is_device_cert_model;
-
-    is_requester_cert = false;
-
-    /*defalut is true*/
-    is_device_cert_model = true;
-
-    *data = NULL;
-    *size = 0;
-    if (hash != NULL) {
-        *hash = NULL;
-    }
-    if (hash_size != NULL) {
-        *hash_size = 0;
-    }
-
-    switch (chain_id) {
-    case LIBSPDM_TEST_CERT_SMALL: /* data_size smaller than 1024 Bytes*/
-        file = "long_chains/Shorter1024B_bundle_responder.certchain.der";
-        break;
-    case LIBSPDM_TEST_CERT_MAXINT16: /* data_size slightly smaller than 0x7FFF*/
-        file = "long_chains/ShorterMAXINT16_bundle_responder.certchain.der";
-        break;
-    case LIBSPDM_TEST_CERT_MAXUINT16: /* data_size slightly smaller than 0xFFFF*/
-        file = "long_chains/ShorterMAXUINT16_bundle_responder.certchain.der";
-        break;
-    case LIBSPDM_LIBSPDM_TEST_CERT_MAXUINT16_LARGER: /* data_size larger than 0xFFFF*/
-        file = "long_chains/LongerMAXUINT16_bundle_responder.certchain.der";
-        break;
-    default:
-        LIBSPDM_ASSERT(false);
-        return false;
-    }
-    res = libspdm_read_input_file(file, &file_data, &file_size);
-    if (!res) {
-        return res;
-    }
-
-    digest_size = libspdm_get_hash_size(base_hash_algo);
-
-    cert_chain_size = sizeof(spdm_cert_chain_t) + digest_size + file_size;
-    cert_chain = (void *)malloc(cert_chain_size);
-    if (cert_chain == NULL) {
-        free(file_data);
-        return false;
-    }
-    cert_chain->length = (uint16_t)cert_chain_size;
-    cert_chain->reserved = 0;
-
-    res = libspdm_verify_cert_chain_data(file_data, file_size,
-                                         base_asym_algo, base_hash_algo,
-                                         is_requester_cert, is_device_cert_model);
-    if (!res) {
-        free(file_data);
-        free(cert_chain);
-        return res;
-    }
-
-
-    /* Get Root Certificate and calculate hash value*/
-
-    res = libspdm_x509_get_cert_from_cert_chain(file_data, file_size, 0, &root_cert,
-                                                &root_cert_len);
-    if (!res) {
-        free(file_data);
-        free(cert_chain);
-        return res;
-    }
-
-    res = libspdm_hash_all(base_hash_algo, root_cert, root_cert_len,
-                           (uint8_t *)(cert_chain + 1));
-    if (!res) {
-        free(file_data);
-        free(cert_chain);
-        return res;
-    }
-    libspdm_copy_mem((uint8_t *)cert_chain + sizeof(spdm_cert_chain_t) + digest_size,
-                     cert_chain_size - (sizeof(spdm_cert_chain_t) + digest_size),
-                     file_data, file_size);
-
-    *data = cert_chain;
-    *size = cert_chain_size;
-    if (hash != NULL) {
-        *hash = (cert_chain + 1);
-    }
-    if (hash_size != NULL) {
-        *hash_size = digest_size;
-    }
-
-    free(file_data);
-    return true;
+    res = libspdm_read_input_file(file, data, size);
+    return res;
 }
